chore: move Label component + remove link from Koeb
